+++ webpack/webpack.web.js
@@ -1,13 +1,13 @@
 'use strict'
 
 const path = require('path')
-const webpackMerge = require('webpack-merge')
+const { merge } = require('webpack-merge')
 
 const commonConfig = require('./webpack.common.js')()
 
 module.exports = function (options) {
   delete commonConfig.externals;
-  return webpackMerge(commonConfig, {
+  return merge(commonConfig, {
     output: {
       libraryTarget: 'umd',
       path: path.join(__dirname, '../dist/web'),